fixed non-critical block/thread count equation issues
Cleaned up some non-critical issues.   The calculations for tpb and bpg were some quick copy-pasta I forgot to double check, because it looked right.  It wasn't but it did not affect the output because they were too large in all cases, and the bounds checking already within the GPU code prevented the extra threads from producing output.  This simply meant we always had extra blocks of threads running than we needed.  Now it should be optimal.....